Create GFF_GeneSearch.py
Searching and returning GeneID and nucleotide sequence coordinates for GFF files. Gene ID is hardcoded and can be changed depending on search criteria. Post processing of the nucleotide sequence is also included, converting all nucleotide sequences on the negative strand to their reverse complement on the positive strand of DNA.